# api/swagger.yaml
basePath: /v1
definitions:
...
  domain.Transaction:
    properties:
      account_id:
        type: string
      amount_cents:
        description: Always in cents; positive value; type determines direction
        type: integer
      category_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      master_purchase_id:
        description: Empty for regular transactions; set in Phase 2
        type: string
      occurred_at:
        type: string
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/domain.TransactionType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.TransactionType:
    enum:
    - income
    - expense
    - transfer
    type: string
    x-enum-varnames:
    - TransactionTypeIncome
    - TransactionTypeExpense
    - TransactionTypeTransfer
...
  handler.CreateTransactionRequest:
    properties:
      account_id:
        type: string
      amount_cents:
        type: integer
      category_id:
        type: string
      description:
        maxLength: 255
        minLength: 1
        type: string
      occurred_at:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.TransactionType'
        enum:
        - income
        - expense
        - transfer
    required:
    - account_id
    - amount_cents
    - category_id
    - description
    - occurred_at
    - type
    type: object
...
  handler.UpdateTransactionRequest:
    properties:
      amount_cents:
        type: integer
      category_id:
        type: string
      description:
        maxLength: 255
        minLength: 1
        type: string
      occurred_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
  description: Moolah Household Finance & Investment SaaS API.
  title: Moolah Financial API
  version: "1.0"
paths:
...
  /v1/transactions:
    get:
      consumes:
      - application/json
      description: Returns a list of transactions for the current tenant. Can be filtered
        by account, category, type, and date range.
      parameters:
      - description: Account ULID
        in: query
        name: account_id
        type: string
      - description: Category ULID
        in: query
        name: category_id
        type: string
      - description: Transaction type (income/expense/transfer)
        in: query
        name: type
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Limit (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Creates a new financial transaction (income, expense, or transfer).
      parameters:
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Transaction details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /v1/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a transaction.
      parameters:
      - description: Transaction ULID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Transaction deleted
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: Returns details of a specific transaction by ID.
      parameters:
      - description: Transaction ULID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Transaction'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get transaction
      tags:
      - transactions
    patch:
      consumes:
      - application/json
      description: Updates details of an existing transaction.
      parameters:
      - description: Transaction ULID
        in: path
        name: id
        required: true
        type: string
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Update fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update transaction
      tags:
      - transactions
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"